// src/blockdb/schema.ts
export interface DefaultCards20260721211623 {
  object: string;
  id: string;
  oracle_id: string;
  multiverse_ids?: unknown;
  mtgo_id: number;
  arena_id: number;
  tcgplayer_id: number;
  cardmarket_id: number;
  name: string;
  lang: string;
  released_at: string;
  uri: string;
  scryfall_uri: string;
  layout: string;
  highres_image: boolean;
  image_status: string;
  image_updated_at: string;
  image_uris?: unknown;
  mana_cost: string;
  cmc: number;
  type_line: string;
  oracle_text: string;
  colors: string[];
  color_identity: string[];
  keywords: string[];
  produced_mana?: unknown;
  all_parts?: unknown;
  legalities?: unknown;
  games: string[];
  reserved: boolean;
  game_changer: boolean;
  foil: boolean;
  nonfoil: boolean;
  finishes?: unknown;
  oversized: boolean;
  promo: boolean;
  reprint: boolean;
  variation: boolean;
  set_id: string;
  set: string;
  set_name: string;
  set_type: string;
  set_uri: string;
  set_search_uri: string;
  scryfall_set_uri: string;
  rulings_uri: string;
  prints_search_uri: string;
  collector_number: string;
  digital: boolean;
  rarity: string;
  card_back_id: string;
  artist: string;
  artist_ids?: unknown;
  illustration_id: string;
  border_color: string;
  frame: string;
  full_art: boolean;
  textless: boolean;
  booster: boolean;
  story_spotlight: boolean;
  prices?: unknown;
  related_uris?: unknown;
  purchase_uris?: unknown;
  mtgo_foil_id: number;
  power: string;
  toughness: string;
  flavor_text: string;
  edhrec_rank: number;
  penny_rank: number;
  promo_types?: unknown;
  security_stamp: string;
  card_faces?: unknown;
  preview?: unknown;
  watermark: string;
  frame_effects?: unknown;
  loyalty: string;
  printed_name: string;
  resource_id: string;
  tcgplayer_etched_id: number;
  flavor_name: string;
  attraction_lights?: unknown;
  color_indicator?: unknown;
  printed_text: string;
  printed_type_line: string;
  power_num?: number;
  toughness_num?: number;
  loyalty_num?: number;
  name_fold?: string;
  type_line_fold?: string;
  oracle_text_fold?: string;
  artist_fold?: string;
  flavor_text_fold?: string;
  set_name_fold?: string;
}

export interface Records {
  "default-cards-20260721211623": DefaultCards20260721211623;
}